+++ src/WebRoutes.tsx
@@ -60,6 +60,9 @@ import ConditionFarmer from "./pages/setting/pointSetting/ConditionFarmer";
 import IndexReceivePoint from "./pages/point/IndexReceivePoint";
 import IndexPlanningPoint from "./pages/point/IndexPlanningPoint";
 import DetailReceivePoint from "./pages/point/DetailReceivePoint";
+import IndexRedeem from "./pages/redeem/IndexRedeem";
+import DetailFarmerRedeem from "./pages/redeem/DetailFarmerRedeem";
+import DetailDronerRedeem from "./pages/redeem/DetailDronerRedeem";
 
 const WebRoutes: React.FC<any> = () => {
   return (
@@ -113,7 +116,6 @@ const WebRoutes: React.FC<any> = () => {
           <Route path="*" element={<PageNotFound />} />
           <Route path="/EditPromotion/:id" element={<EditPromotion />} />
           <Route path="*" element={<PageNotFound />} />
-          IndexReport
           <Route path="/AddPromotion" element={<AddPromotion />} />
           <Route path="/AddNews" element={<AddNewsPage />} />
           <Route path="/IndexReport" element={<IndexReport />} />
@@ -150,8 +152,16 @@ const WebRoutes: React.FC<any> = () => {
             path="/DetailReceivePoint/:id"
             element={<DetailReceivePoint />}
           />
+          <Route path="/IndexRedeem/:type" element={<IndexRedeem />} />
+          <Route
+            path="/DetailFarmerRedeem/:id"
+            element={<DetailFarmerRedeem />}
+          />
+          <Route
+            path="/DetailDronerRedeem/:id"
+            element={<DetailDronerRedeem />}
+          />
         </Route>
-
         <Route element={<PublicRoute />}>
           <Route index element={<AuthPage />} />
           <Route path="/ErrorLoginPage" element={<ErrorLoginPage />} />