+++ src/pages/campaign/point/EditCampaignPoint.tsx
@@ -175,27 +175,27 @@ const EditCampaignPoint = () => {
         <Row>
           <BackIconButton onClick={() => navigate(-1)} />
           <span className="pt-3">
-            <strong style={{ fontSize: "20px" }}>แก้ไขแคมเปญคะแนน</strong>
+            <strong style={{ fontSize: "20px" }}>แก้ไขแคมเปญแต้ม</strong>
           </span>
         </Row>
         <CardContainer>
-          <CardHeader textHeader="ข้อมูลแคมเปญคะแนน" />
+          <CardHeader textHeader="ข้อมูลแคมเปญแต้ม" />
           <Form style={{ padding: "32px" }} form={form}>
             <Col span={24}>
               <label>
-                ชื่อแคมเปญคะแนน<span style={{ color: color.Error }}>*</span>
+                ชื่อแคมเปญแต้ม<span style={{ color: color.Error }}>*</span>
               </label>
               <Form.Item
                 name="campaignName"
                 rules={[
                   {
                     required: true,
-                    message: "กรุณากรอกชื่อแคมเปญคะแนน!",
+                    message: "กรุณากรอกชื่อแคมเปญแต้ม!",
                   },
                 ]}
               >
                 <Input
-                  placeholder="กรอกชื่อแคมเปญคะแนน"
+                  placeholder="กรอกชื่อแคมเปญแต้ม"
                   autoComplete="off"
                   disabled={isEdit}
                 />
@@ -316,20 +316,20 @@ const EditCampaignPoint = () => {
             <Row gutter={8} justify={"start"}>
               <Col span={7}>
                 <label>
-                  คะแนนที่ได้รับ<span style={{ color: color.Error }}>*</span>
+                  แต้มที่ได้รับ<span style={{ color: color.Error }}>*</span>
                 </label>
                 <Form.Item
                   name="point"
                   rules={[
                     {
                       required: true,
-                      message: "กรุณากรอกจำนวนคะแนน!",
+                      message: "กรุณากรอกจำนวนแต้ม!",
                     },
                   ]}
                 >
                   <Input
-                    placeholder="กรอกคะแนนที่ได้รับ"
-                    suffix="คะแนน"
+                    placeholder="กรอกแต้มที่ได้รับ"
+                    suffix="แต้ม"
                     autoComplete="off"
                     disabled={isActive}
                     onChange={(e) => checkNumber(e, "point")}
@@ -426,7 +426,7 @@ const EditCampaignPoint = () => {
               โปรดตรวจสอบแคมเปญที่คุณต้องการแก้ไข ก่อนที่จะกดยืนยันแก้ไข
             </span>
             <p className="text-secondary">
-              เพราะอาจส่งผลต่อการคะแนนในแอปพลิเคชัน
+              เพราะอาจส่งผลต่อการแต้มในแอปพลิเคชัน
             </p>
           </div>
           <Divider