+++ src/pages/point/DetailReceivePoint.tsx
@@ -6,7 +6,6 @@ import { useNavigate } from "react-router-dom";
 import { BackIconButton } from "../../components/button/BackButton";
 import { CardContainer } from "../../components/card/CardContainer";
 import { CardHeader } from "../../components/header/CardHearder";
-import Layouts from "../../components/layout/Layout";
 import { color } from "../../resource";
 import { DateTimeUtil } from "../../utilities/DateTimeUtil";
 import { numberWithCommasToFixed } from "../../utilities/TextFormatter";
@@ -64,6 +63,7 @@ const DetailDronerPoint = () => {
       updateBy: "รชยา ก้าวแปดเมตร",
     },
   ];
+  
   const renderDetailRecipient = () => {
     let checkFarmer: any = dataMock.description.filter(
       (x: any) => x.source === "Farmer"
@@ -361,7 +361,7 @@ const DetailDronerPoint = () => {
   };
 
   return (
-    <Layouts>
+    <>
       <Row>
         <BackIconButton onClick={() => navigate(-1)} />
         <span className="pt-3">
@@ -373,7 +373,7 @@ const DetailDronerPoint = () => {
       {renderDetailTask()}
       <br />
       {renderHistory()}
-    </Layouts>
+    </>
   );
 };