+++ src/pages/task/newTask/IndexNewTask.tsx
@@ -3,17 +3,14 @@ import {
   EditOutlined,
   UserOutlined,
   ExceptionOutlined,
-  InfoCircleFilled,
 } from "@ant-design/icons";
 import {
   Badge,
   Button,
   DatePicker,
-  Divider,
   Dropdown,
   Menu,
   Pagination,
-  Popover,
   Select,
   Table,
 } from "antd";
@@ -41,7 +38,6 @@ import { color } from "../../../resource";
 import { DateTimeUtil } from "../../../utilities/DateTimeUtil";
 import {
   numberWithCommas,
-  numberWithCommasToFixed,
 } from "../../../utilities/TextFormatter";
 import { DashboardLayout } from "../../../components/layout/Layout";
 import { useNavigate } from "react-router-dom";
@@ -74,7 +70,6 @@ const IndexNewTask = () => {
       searchStartDate,
       searchEndDate
     ).then((res) => {
-      console.log(res.data);
       setData(res);
     });
   };