+++ src/pages/task/todayTask/DetailEdit/EditWaitStart.tsx
@@ -554,10 +554,7 @@ function EditWaitStart() {
           <div className="col-lg-4 text-start">
             <label>จำนวนไร่</label>
             <Form.Item>
-              <Input
-                defaultValue={data.farmAreaAmount}
-                suffix="ไร่"
-              />
+              <Input defaultValue={data.farmAreaAmount} suffix="ไร่" disabled/>
             </Form.Item>
           </div>
         </div>
@@ -768,6 +765,26 @@ function EditWaitStart() {
         </div>
         <div className="row pt-3">
           <div className="form-group col-lg-6 p-2">
+            <label>จำนวนแต้มที่ใช้แลก</label>
+            <Input
+              suffix="แต้ม"
+              value={data.usePoint}
+              disabled
+              autoComplete="off"
+            />
+          </div>
+          <div className="form-group col-lg-6 p-2">
+            <label>ส่วนลดจากการใช้แต้ม</label>
+            <Input
+              suffix="บาท"
+              value={data.discountCampaignPoint}
+              disabled
+              autoComplete="off"
+            />
+          </div>
+        </div>
+        {/* <div className="row pt-3">
+          <div className="form-group col-lg-6 p-2">
             <label>โปรโมชั่นนักบินโดรน</label>
             <Input
               suffix="บาท"
@@ -785,7 +802,7 @@ function EditWaitStart() {
               autoComplete="off"
             />
           </div>
-        </div>
+        </div> */}
       </Form>
     </Form>
   );